refactor: Created sendEmail.ts
