+++ src/server/lib/trackMetric.ts
@@ -49,37 +49,3 @@ export const trackMetric = (
       logger.error(error.message);
     });
 };
-
-const SESV2 = new AWS.SESV2(standardAwsConfig);
-
-export const sendEmail = (
-  html: string,
-  plainText: string,
-  subject: string,
-  toAddress: string,
-  fromAddress: string,
-) => {
-  const params: AWS.SESV2.SendEmailRequest = {
-    Content: {
-      Simple: {
-        Body: {
-          Html: {
-            Data: html,
-          },
-          Text: {
-            Data: plainText,
-          },
-        },
-        Subject: {
-          Data: subject,
-        },
-      },
-    },
-    Destination: {
-      ToAddresses: [toAddress],
-    },
-    FromEmailAddress: fromAddress,
-  };
-
-  return SESV2.sendEmail(params).promise();
-};
